Lag test for feil aarsak til innsending
