+++ src/test/kotlin/no/nav/helsearbeidsgiver/authorization/AuthApiTest.kt
@@ -9,18 +9,10 @@ import io.ktor.client.request.setBody
 import io.ktor.http.ContentType
 import io.ktor.http.HttpStatusCode
 import io.ktor.http.contentType
-import io.mockk.Runs
 import io.mockk.every
-import io.mockk.just
-import io.mockk.mockkStatic
-import io.mockk.unmockkStatic
-import io.mockk.verify
 import kotlinx.coroutines.test.runTest
-import no.nav.helsearbeidsgiver.config.Services
-import no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.AarsakInnsending
 import no.nav.helsearbeidsgiver.forespoersel.ForespoerselResponse
 import no.nav.helsearbeidsgiver.forespoersel.Status
-import no.nav.helsearbeidsgiver.inntektsmelding.Arbeidsgiver
 import no.nav.helsearbeidsgiver.inntektsmelding.InntektsmeldingFilterResponse
 import no.nav.helsearbeidsgiver.inntektsmelding.InntektsmeldingRequest
 import no.nav.helsearbeidsgiver.utils.DEFAULT_ORG
@@ -30,10 +22,7 @@ import no.nav.helsearbeidsgiver.utils.json.toJson
 import no.nav.helsearbeidsgiver.utils.mockForespoersel
 import no.nav.helsearbeidsgiver.utils.mockInntektsmeldingRequest
 import no.nav.helsearbeidsgiver.utils.mockInntektsmeldingResponse
-import no.nav.helsearbeidsgiver.utils.opprettImTransaction
-import no.nav.helsearbeidsgiver.utils.test.wrapper.genererGyldig
 import no.nav.helsearbeidsgiver.utils.ugyldigTokenManglerSystembruker
-import no.nav.helsearbeidsgiver.utils.wrapper.Orgnr
 import org.junit.jupiter.api.Test
 import java.util.UUID
 
@@ -116,127 +105,5 @@ class AuthApiTest : ApiTest() {
             val inntektsmeldingFilterResponse = response.body<InntektsmeldingFilterResponse>()
             inntektsmeldingFilterResponse.antall shouldBe 1
             inntektsmeldingFilterResponse.inntektsmeldinger[0].arbeidsgiver.orgnr shouldBe DEFAULT_ORG
-        }
-
-    @Test
-    fun `innsending av inntektsmelding på gyldig forespørsel`() =
-        runTest {
-            // Nødvendig pga transaction rundt service kall
-            mockkStatic(Services::opprettImTransaction)
-            every { services.opprettImTransaction(any(), any()) } just Runs
-            val requestBody = mockInntektsmeldingRequest()
-            val forespoersel = mockForespoersel().copy(navReferanseId = requestBody.navReferanseId, orgnr = DEFAULT_ORG)
-            every { repositories.forespoerselRepository.hentForespoersel(forespoersel.navReferanseId) } returns forespoersel
-            every { repositories.inntektsmeldingRepository.hent(forespoersel.navReferanseId) } returns emptyList()
-            val response =
-                client.post("/v1/inntektsmelding") {
-                    bearerAuth(mockOAuth2Server.gyldigSystembrukerAuthToken(DEFAULT_ORG))
-                    contentType(ContentType.Application.Json)
-                    setBody(requestBody.toJson(serializer = InntektsmeldingRequest.serializer()))
-                }
-            response.status shouldBe HttpStatusCode.Created
-            verify {
-                services.opprettImTransaction(
-                    match { it.type.id == requestBody.navReferanseId },
-                    match { it.type.id == requestBody.navReferanseId },
-                )
-            }
-            unmockkStatic(Services::opprettImTransaction)
-        }
-
-    @Test
-    fun `innsending av inntektsmelding uten gyldig forespørsel gir bad request`() =
-        runTest {
-            // Nødvendig pga transaction rundt service kall
-            mockkStatic(Services::opprettImTransaction)
-            every { services.opprettImTransaction(any(), any()) } just Runs
-            val requestBody = mockInntektsmeldingRequest()
-            every { repositories.forespoerselRepository.hentForespoersel(requestBody.navReferanseId) } returns null
-            val response =
-                client.post("/v1/inntektsmelding") {
-                    bearerAuth(mockOAuth2Server.gyldigSystembrukerAuthToken(DEFAULT_ORG))
-                    contentType(ContentType.Application.Json)
-                    setBody(requestBody.toJson(serializer = InntektsmeldingRequest.serializer()))
-                }
-            response.status shouldBe HttpStatusCode.BadRequest
-            verify(exactly = 0) {
-                services.opprettImTransaction(
-                    match { it.type.id == requestBody.navReferanseId },
-                    match { it.type.id == requestBody.navReferanseId },
-                )
-            }
-            unmockkStatic(Services::opprettImTransaction)
-        }
-
-    @Test
-    fun `innsending av duplikat inntektsmelding gyldig forespørsel gir conflict`() =
-        runTest {
-            // Nødvendig pga transaction rundt service kall
-            mockkStatic(Services::opprettImTransaction)
-            every { services.opprettImTransaction(any(), any()) } just Runs
-            val requestBody = mockInntektsmeldingRequest().copy(aarsakInnsending = AarsakInnsending.Endring)
-            val forespoersel = mockForespoersel().copy(navReferanseId = requestBody.navReferanseId, orgnr = DEFAULT_ORG)
-            every { repositories.forespoerselRepository.hentForespoersel(forespoersel.navReferanseId) } returns forespoersel
-            every { repositories.inntektsmeldingRepository.hent(forespoersel.navReferanseId) } returns
-                listOf(
-                    mockInntektsmeldingResponse().copy(
-                        id = UUID.randomUUID(),
-                        navReferanseId = requestBody.navReferanseId,
-                        arbeidsgiver = Arbeidsgiver(DEFAULT_ORG, requestBody.arbeidsgiverTlf),
-                        inntekt = requestBody.inntekt,
-                        refusjon = requestBody.refusjon,
-                        agp = requestBody.agp,
-                    ),
-                )
-            val im = buildInntektsmelding(forespoerselId = forespoersel.navReferanseId)
-            every { repositories.inntektsmeldingRepository.hent(DEFAULT_ORG) } returns
-                listOf(mockInntektsmeldingResponse(im))
-            val response =
-                client.post("/v1/inntektsmelding") {
-                    bearerAuth(mockOAuth2Server.gyldigSystembrukerAuthToken(DEFAULT_ORG))
-                    contentType(ContentType.Application.Json)
-                    setBody(requestBody.toJson(serializer = InntektsmeldingRequest.serializer()))
-                }
-            response.status shouldBe HttpStatusCode.Conflict
-            verify(exactly = 0) {
-                services.opprettImTransaction(
-                    match { it.type.id == requestBody.navReferanseId },
-                    match { it.type.id == requestBody.navReferanseId },
-                )
-            }
-            unmockkStatic(Services::opprettImTransaction)
-        }
-
-    @Test
-    fun `innsending av inntektsmelding på feil orgnr gir feil`() =
-        runTest {
-            val requestBody = mockInntektsmeldingRequest()
-            val forespoersel =
-                mockForespoersel().copy(
-                    navReferanseId = requestBody.navReferanseId,
-                    orgnr = Orgnr.genererGyldig().verdi,
-                )
-            every { repositories.forespoerselRepository.hentForespoersel(forespoersel.navReferanseId) } returns forespoersel
-            val response =
-                client.post("/v1/inntektsmelding") {
-                    bearerAuth(mockOAuth2Server.gyldigSystembrukerAuthToken(DEFAULT_ORG))
-                    contentType(ContentType.Application.Json)
-                    setBody(requestBody.toJson(serializer = InntektsmeldingRequest.serializer()))
-                }
-            response.status shouldBe HttpStatusCode.BadRequest
-        }
-
-    @Test
-    fun `innsending av inntektsmelding på forespoersel som ikke finnes gir feil`() =
-        runTest {
-            val requestBody = mockInntektsmeldingRequest()
-            every { repositories.forespoerselRepository.hentForespoersel(requestBody.navReferanseId) } returns null
-            val response =
-                client.post("/v1/inntektsmelding") {
-                    bearerAuth(mockOAuth2Server.gyldigSystembrukerAuthToken(DEFAULT_ORG))
-                    contentType(ContentType.Application.Json)
-                    setBody(requestBody.toJson(serializer = InntektsmeldingRequest.serializer()))
-                }
-            response.status shouldBe HttpStatusCode.BadRequest
         }
 }
